no-overlay: Extend AddressSet API to expose UUID
Add GetASUUID() method to the AddressSet interface and its
implementations (ovnAddressSets and fakeAddressSets) to expose
the UUIDs of IPv4 and IPv6 address sets.

This is needed for the no-overlay mode SNAT exemption feature
(OKEP-5259), where the UUID is used to reference address sets
in the exempted_ext_ips field of OVN SNAT rules. Using the UUID
directly (instead of the hash name) allows OVN to efficiently
match and exempt traffic destined to pod subnets from being SNATed.

The method returns a tuple of (v4UUID, v6UUID) strings, with empty
strings for address sets that don't exist.

Related changes:
- Add GetASUUID() to AddressSet interface
- Implement GetASUUID() in ovnAddressSets
- Implement GetASUUID() in fakeAddressSets for testing
- Update mocks to include the new method

Signed-off-by: Peng Liu <pliu@redhat.com>